PR115: addressed some style issues. updated pipeline LICENSE.md
