+++ specs/lib/triangleAxisAlignedBoundingBoxOverlapSpec.js
@@ -1,9 +1,10 @@
 'use strict';
 var Cesium = require('cesium');
+var AxisAlignedBoundingBox = Cesium.AxisAlignedBoundingBox;
 var Cartesian3 = Cesium.Cartesian3;
-var triBoxOverlap = require('../../lib/triBoxOverlap');
+var triangleAxisAlignedBoundingBoxOverlap = require('../../lib/triangleAxisAlignedBoundingBoxOverlap');
 
-describe('triBoxOverlap', function() {
+describe('triangleAxisAlignedBoundingBoxOverlap', function() {
     var triangle = [
         new Cartesian3(0.0, 0.0, 0.0),
         new Cartesian3(1.0, 0.0, 0.0),
@@ -13,6 +14,14 @@ describe('triBoxOverlap', function() {
     var boxCenter = new Cartesian3();
     var halfSize = new Cartesian3();
 
+    var boundingBoxScratch = new AxisAlignedBoundingBox();
+    function makeAxisAlignedBoundingBox(center, halfSize) {
+        boundingBoxScratch.center = center;
+        boundingBoxScratch.maximum = Cartesian3.add(center, halfSize, boundingBoxScratch.maximum);
+        boundingBoxScratch.minimum = Cartesian3.subtract(center, halfSize, boundingBoxScratch.minimum);
+        return boundingBoxScratch;
+    }
+
     it('correctly detects no intersection for a completely separate triangle and box', function() {
         boxCenter.x = 2.0;
         boxCenter.y = 2.0;
@@ -21,7 +30,9 @@ describe('triBoxOverlap', function() {
         halfSize.y = 0.5;
         halfSize.z = 0.5;
 
-        expect(triBoxOverlap(boxCenter, halfSize, triangle)).toEqual(false);
+        var boundingBox = makeAxisAlignedBoundingBox(boxCenter, halfSize);
+
+        expect(triangleAxisAlignedBoundingBoxOverlap(boundingBox, triangle)).toEqual(false);
     });
 
     it('correctly detects an intersection when the triangle is entirely inside the box', function() {
@@ -32,7 +43,9 @@ describe('triBoxOverlap', function() {
         halfSize.y = 1.0;
         halfSize.z = 1.0;
 
-        expect(triBoxOverlap(boxCenter, halfSize, triangle)).toEqual(true);
+        var boundingBox = makeAxisAlignedBoundingBox(boxCenter, halfSize);
+
+        expect(triangleAxisAlignedBoundingBoxOverlap(boundingBox, triangle)).toEqual(true);
     });
 
     it('correctly detects an intersection when the box is entirely in the triangle plane', function() {
@@ -43,7 +56,9 @@ describe('triBoxOverlap', function() {
         halfSize.y = 0.1;
         halfSize.z = 0.1;
 
-        expect(triBoxOverlap(boxCenter, halfSize, triangle)).toEqual(true);
+        var boundingBox = makeAxisAlignedBoundingBox(boxCenter, halfSize);
+
+        expect(triangleAxisAlignedBoundingBoxOverlap(boundingBox, triangle)).toEqual(true);
     });
 
     it('correctly detects an intersection when a triangle vertex is in the box', function() {
@@ -54,7 +69,9 @@ describe('triBoxOverlap', function() {
         halfSize.y = 0.5;
         halfSize.z = 0.5;
 
-        expect(triBoxOverlap(boxCenter, halfSize, triangle)).toEqual(true);
+        var boundingBox = makeAxisAlignedBoundingBox(boxCenter, halfSize);
+
+        expect(triangleAxisAlignedBoundingBoxOverlap(boundingBox, triangle)).toEqual(true);
     });
 
     it('correctly detects an intersection when the intersection is on an edge of the triangle', function() {
@@ -65,6 +82,8 @@ describe('triBoxOverlap', function() {
         halfSize.y = 1.1;
         halfSize.z = 1.1;
 
-        expect(triBoxOverlap(boxCenter, halfSize, triangle)).toEqual(true);
+        var boundingBox = makeAxisAlignedBoundingBox(boxCenter, halfSize);
+
+        expect(triangleAxisAlignedBoundingBoxOverlap(boundingBox, triangle)).toEqual(true);
     });
 });
